+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
 ### Added
 - Endpoints to return allowed demand types and units [#155](https://github.com/IN-CORE/incore-services/issues/155)
 
@@ -15,7 +16,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Upgraded project and semantics-service dependencies: Jersey 3.1.2 and swagger 2.1.12 for OpenAPI 3 [#167](https://github.com/IN-CORE/incore-services/issues/167)
 - Upgraded project and hazard-service dependencies: Jersey 3.1.2 and swagger 2.1.12 for OpenAPI 3 [#163](https://github.com/IN-CORE/incore-services/issues/163)
 - Upgraded project and data-service dependencies: Jersey 3.1.2 and swagger 2.1.12 for OpenAPI 3 [#165](https://github.com/IN-CORE/incore-services/issues/165)
+- Refactor DAO for Semantics service [#151](https://github.com/IN-CORE/incore-services/issues/151)
 - Removed jetty-runner 9 and replaced with Jetty 11 docker image [#174](https://github.com/IN-CORE/incore-services/issues/174)
+- Removed maestro service from the incore services repository [#145](https://github.com/IN-CORE/incore-services/issues/145)
+
+### Fixed
+- Missing log4j.properties file required to configure logging [#148](https://github.com/IN-CORE/incore-services/issues/148)
 
 ## [1.14.0] -2023-06-14
 ### Added